Make sure that cache.js has a non-zero size
If the `node` command here ever fails, `cache.js` will wind up existing
but being a useless zero-byte file.

Switch from `-f` (which tests for existence only) to `-s` (which
requires that the file be non-zero in size) to make sure that we re-try
the download the next time we run the script in that case.

I believe this is a POSIX standard switch, so should work with pretty
much any shell pretending to be Bash.